Update to v2.4.11

+++ tools/deps/install_boost.py
@@ -115,8 +115,9 @@ def InstallDependency():
   for af in aflist:
     msg.HeadPrint("Ready: \"{}\"".format(af))
   # Cleanup.
-  msg.Print("Cleaning temporary directory...")
-  utl.RemoveDirectory(temp_dir)
+  if "--no-clean" not in sys.argv:
+    msg.Print("Cleaning temporary directory...")
+    utl.RemoveDirectory(temp_dir)
 
 
 if __name__ == "__main__":
